[amd64] enable CONFIG_DRM_SIMPLEDRM and CONFIG_DRM as built-in, re-enable CONFIG_SYSFB_SIMPLEFB
This enables UEFI systems to use the EFI framebuffer, so that a graphical
Plymouth output can be displayed immediately, without requiring hardware
specific kernel drivers and firmware.

This is especially useful to build generic UKIs that still
support Plymouth in the initrd by adding 'plymouth.use-simpledrm' to
the kernel command line.

Fedora 42 switched to this for their builds too.

https://fedoraproject.org/wiki/Changes/PlymouthUseSimpledrm
https://gitlab.freedesktop.org/plymouth/plymouth/-/issues/264

By enabling these as built-ins the setup in the initrd is much simpler,
and it matches how other distros (Fedora, Ubuntu) use this functionality.

CONFIG_SYSFB_SIMPLEFB was disabled for #822575 as the simpledrm driver
didn't exist at the time. The kernel maintainer specifically said
to disable it until simpledrm was available.

It is now available and enabled, so turn this option back on so
that it can be used with the EFI framebuffer.

Closes: #993640