[arm64,ppc64*] Disable broken graphics drivers when using non-4 kiB pages
nouveau is long known to only work with 4 kiB pages (bug #790690), but
lacks a Kconfig dependency to enforce this.

xe is also known to only work with 4 kiB pages (bug #1109776) and
again lacks a Kconfig dependency.  There are patches that may fix this
soon, but they have not yet landed.

Explicitly disable both of these in the config fragments that set
different page sizes.